--- a/.gitignore
+++ b/.gitignore
@@ -10,6 +10,9 @@ junit.xml
 coverage/
 .docz/
 *.js.map
+.yarn
+platform/docs/.docusaurus
+platform/ui/storybook-static
 
 # YALC (for Erik)
 .yalc