--- a/extensions/cornerstone/package.json
+++ b/extensions/cornerstone/package.json
@@ -35,7 +35,7 @@
     "cornerstone-core": "^2.2.8",
     "cornerstone-math": "^0.1.8",
     "cornerstone-tools": "^4.20.1",
-    "cornerstone-wado-image-loader": "^3.1.0",
+    "cornerstone-wado-image-loader": "^4.0.0-rc1",
     "dcmjs": "0.18.8",
     "dicom-parser": "^1.8.3",
     "hammerjs": "^2.0.8",