--- a/extensions/cornerstone/package.json
+++ b/extensions/cornerstone/package.json
@@ -32,12 +32,12 @@
   "peerDependencies": {
     "@ohif/core": "^0.50.0",
     "@ohif/ui": "^0.50.0",
-    "cornerstone-core": "^2.2.8",
-    "cornerstone-math": "^0.1.8",
-    "cornerstone-tools": "^4.20.1",
-    "cornerstone-wado-image-loader": "^4.0.0-rc1",
+    "cornerstone-core": "^2.5.0",
+    "cornerstone-math": "^0.1.9",
+    "cornerstone-tools": "^6.0.0",
+    "cornerstone-wado-image-loader": "^4.0.0",
     "dcmjs": "0.18.8",
-    "dicom-parser": "^1.8.3",
+    "dicom-parser": "^1.8.8",
     "hammerjs": "^2.0.8",
     "prop-types": "^15.6.2",
     "react": "^16.8.6",
@@ -51,6 +51,6 @@
     "classnames": "^2.2.6",
     "lodash.merge": "^4.6.2",
     "lodash.throttle": "^4.1.1",
-    "react-cornerstone-viewport": "4.0.5"
+    "react-cornerstone-viewport": "4.1.1"
   }
 }