--- a/extensions/cornerstone/package.json
+++ b/extensions/cornerstone/package.json
@@ -35,7 +35,7 @@
     "cornerstone-core": "^2.5.0",
     "cornerstone-math": "^0.1.9",
     "cornerstone-tools": "^6.0.0",
-    "cornerstone-wado-image-loader": "^4.0.2",
+    "cornerstone-wado-image-loader": "^4.0.3-rc1",
     "dcmjs": "0.18.8",
     "dicom-parser": "^1.8.9",
     "hammerjs": "^2.0.8",