--- a/extensions/debugging/package.json
+++ b/extensions/debugging/package.json
@@ -28,7 +28,7 @@
   },
   "peerDependencies": {
     "@ohif/core": "^2.6.0",
-    "dicom-parser": "^1.8.3"
+    "dicom-parser": "^1.8.8"
   },
   "dependencies": {
     "@babel/runtime": "^7.5.5",