--- a/extensions/dicom-segmentation/package.json
+++ b/extensions/dicom-segmentation/package.json
@@ -29,8 +29,8 @@
   },
   "peerDependencies": {
     "@ohif/core": "^0.50.0",
-    "cornerstone-core": "^2.2.8",
-    "cornerstone-tools": "^4.20.1",
+    "cornerstone-core": "^2.5.0",
+    "cornerstone-tools": "^6.0.0",
     "dcmjs": "0.18.8",
     "prop-types": "^15.6.2",
     "react": "^16.8.6",