--- a/extensions/vtk/package.json
+++ b/extensions/vtk/package.json
@@ -50,7 +50,7 @@
   "dependencies": {
     "@babel/runtime": "^7.5.5",
     "lodash.throttle": "^4.1.1",
-    "react-vtkjs-viewport": "^0.14.2"
+    "react-vtkjs-viewport": "^0.14.4"
   },
   "devDependencies": {
     "@ohif/core": "^2.14.1",