--- a/platform/core/package.json
+++ b/platform/core/package.json
@@ -33,7 +33,7 @@
   "peerDependencies": {
     "cornerstone-core": "^2.2.8",
     "cornerstone-tools": "^4.20.1",
-    "cornerstone-wado-image-loader": "^3.1.0",
+    "cornerstone-wado-image-loader": "^4.0.0-rc1",
     "dicom-parser": "^1.8.3"
   },
   "dependencies": {