--- a/platform/core/package.json
+++ b/platform/core/package.json
@@ -31,10 +31,10 @@
     "test:unit:ci": "jest --ci --runInBand --collectCoverage"
   },
   "peerDependencies": {
-    "cornerstone-core": "^2.2.8",
-    "cornerstone-tools": "^4.20.1",
-    "cornerstone-wado-image-loader": "^4.0.0-rc1",
-    "dicom-parser": "^1.8.3"
+    "cornerstone-core": "^2.5.0",
+    "cornerstone-tools": "^6.0.0",
+    "cornerstone-wado-image-loader": "^4.0.0",
+    "dicom-parser": "^1.8.8"
   },
   "dependencies": {
     "@babel/runtime": "^7.5.5",