--- a/platform/core/package.json
+++ b/platform/core/package.json
@@ -33,8 +33,8 @@
   "peerDependencies": {
     "cornerstone-core": "^2.5.0",
     "cornerstone-tools": "^6.0.0",
-    "cornerstone-wado-image-loader": "^4.0.0",
-    "dicom-parser": "^1.8.8"
+    "cornerstone-wado-image-loader": "^4.0.2",
+    "dicom-parser": "^1.8.9"
   },
   "dependencies": {
     "@babel/runtime": "^7.5.5",