--- a/platform/viewer/cypress/integration/common/OHIFUserPreferences.spec.js
+++ b/platform/viewer/cypress/integration/common/OHIFUserPreferences.spec.js
@@ -333,6 +333,13 @@ describe('OHIF User Preferences', () => {
         .scrollIntoView()
         .click();
 
+      // Close Success Message overlay (if displayed)
+      cy.get('body').then(body => {
+        if (body.find('.sb-closeIcon').length > 0) {
+          cy.get('.sb-closeIcon').click({ force: true });
+        }
+      });
+
       // Header should be in "English (USA)""
       cy.get('.research-use')
         .scrollIntoView()
@@ -560,6 +567,13 @@ describe('OHIF User Preferences', () => {
         .scrollIntoView()
         .click();
 
+      // Close Success Message overlay (if displayed)
+      cy.get('body').then(body => {
+        if (body.find('.sb-closeIcon').length > 0) {
+          cy.get('.sb-closeIcon').click({ force: true });
+        }
+      });
+
       // Open User Preferences modal again
       cy.openPreferences();
 
@@ -668,6 +682,13 @@ describe('OHIF User Preferences', () => {
 
       // Save changes
       cy.get('@saveBtn').click();
+      // Close Success Message overlay (if displayed)
+      cy.get('body').then(body => {
+        if (body.find('.sb-closeIcon').length > 0) {
+          cy.get('.sb-closeIcon').click({ force: true });
+        }
+      });
+
       // Open User Preferences modal
       cy.openPreferences();
       // Navigate to Window Level tab