--- a/platform/viewer/package.json
+++ b/platform/viewer/package.json
@@ -65,7 +65,7 @@
     "cornerstone-core": "^2.5.0",
     "cornerstone-math": "^0.1.9",
     "cornerstone-tools": "^6.0.0",
-    "cornerstone-wado-image-loader": "^4.0.3-rc1",
+    "cornerstone-wado-image-loader": "^4.0.3",
     "dcmjs": "0.18.8",
     "dicom-parser": "^1.8.9",
     "dicomweb-client": "^0.8.1",