### PR Checklist

- [ ] Brief description of changes
- [ ] Links to any relevant issues
- [ ] Required status checks are passing
- [ ] User cases if changes impact the user's experience
- [ ] `@mention` a maintainer to request a review

<!--
  Links
  -->

<!-- prettier-ignore-start -->
[blog]: https://circleci.com/blog/triggering-trusted-ci-jobs-on-untrusted-forks/
[script]: https://github.com/jklukas/git-push-fork-to-upstream-branch
<!-- prettier-ignore-end -->




Master branch;

```
@ohif/viewer: This can impact web performance.
@ohif/viewer: Assets:
@ohif/viewer:   app.bundle.3511d353915ee7a8d3f7.js (7.76 MiB)
@ohif/viewer:   vendors~dicom-microscopy-viewer.bundle.6df178d338b4c83f7db4.js (733 KiB)
@ohif/viewer:   11.bundle.762a6624ba301e0cf61b.js (1.09 MiB)
@ohif/viewer: WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
@ohif/viewer: Entrypoints:
@ohif/viewer:   app (7.89 MiB)
@ohif/viewer:       app.8178fc8fd2acb328b940.css
@ohif/viewer:       app.bundle.3511d353915ee7a8d3f7.js
@ohif/viewer:
lerna success run Ran npm script 'build:viewer' in 1 package in 479.3s:
lerna success - @ohif/viewer
✨  Done in 479.95s.
```

This branch:

```
@ohif/viewer: Assets:
@ohif/viewer:   app.bundle.5ba44f148825bbad2628.js (686 KiB)
@ohif/viewer:   vendors~app.bundle.4b804e0cdf28fd79a42e.js (6.9 MiB)
@ohif/viewer:   vendors~dicom-microscopy-viewer.bundle.bae6d317b41d6d6e9eef.js (733 KiB)
@ohif/viewer:   12.bundle.394afb1a7f92cfe0823c.js (1.07 MiB)
@ohif/viewer:   cornerstoneWADOImageLoader.min.js.map (442 KiB)
@ohif/viewer:   index.worker.min.worker.js.map (626 KiB)
@ohif/viewer: WARNING in entrypoint size limit: The following entrypoint(s) combined asset size exceeds the recommended limit (244 KiB). This can impact web performance.
@ohif/viewer: Entrypoints:
@ohif/viewer:   app (7.7 MiB)
@ohif/viewer:       10.ba222fa2e26a9c4c1225.css
@ohif/viewer:       vendors~app.bundle.4b804e0cdf28fd79a42e.js
@ohif/viewer:       app.ba222fa2e26a9c4c1225.css
@ohif/viewer:       app.bundle.5ba44f148825bbad2628.js
@ohif/viewer:
lerna success run Ran npm script 'build:viewer' in 1 package in 103.0s:
lerna success - @ohif/viewer
✨  Done in 104.18s.
```